+++ shaak/modules/word_watch.py
@@ -17,6 +17,7 @@ along with this program. If not, see <http://www.gnu.org/licenses/>.
 '''
 
 # pylint: disable=unsubscriptable-object # pylint/issues/3882
+import asyncio
 import re
 import string
 from dataclasses import dataclass
@@ -34,7 +35,8 @@ from shaak.errors      import InvalidId
 from shaak.helpers     import (MentionType, between_segments, bool2str, commas,
                                get_int_ranges, getrange_s, id2mention,
                                link_to_message, mention2id, pluralize,
-                               resolve_mention, possesivize, chunks, str2bool)
+                               resolve_mention, possesivize, chunks, str2bool,
+                               DiscardingQueue)
 from shaak.matcher     import pattern_preprocess, text_preprocess, word_matches
 from shaak.settings    import product_settings
 from shaak.utils       import ResponseLevel, Utils
@@ -62,6 +64,7 @@ class WordWatch(BaseModule):
         super().__init__(*args, **kwargs)
 
         self.watch_cache: Dict[str, WatchCacheEntry] = {}
+        self.scan_queue = DiscardingQueue(0x400)
         self.bot.add_on_error_hooks(self.after_invoke_hook)
     
     async def add_to_cache(self, watch: WordWatchWatch) -> WatchCacheEntry:
@@ -92,6 +95,8 @@ class WordWatch(BaseModule):
         
         for watch in await WordWatchWatch.all().prefetch_related('guild', 'group'):
             await self.add_to_cache(watch)
+
+        self.bot.loop.create_task(self.scan_loop())
 
         await super().initialize()
     
@@ -130,7 +135,11 @@ class WordWatch(BaseModule):
         if message.channel.category_id and await self.is_id_ignored(message.guild.id, message.channel.category_id):
             return
         
+        if await self.is_id_ignored(message.guild.id, message.author.id):
+            return
+        
         # lots of debug code here. hopefully i can get an error to throw with this better info
+        # 12/5 update: still nothing
         dbg = 'a'
         check_member = message.webhook_id == None
         if isinstance(message.author, discord.User):
@@ -268,10 +277,14 @@ class WordWatch(BaseModule):
                     )
                     fallback_embed.set_footer(text=f'Fallback embed • Ping {product_settings.author_name}!')
                     await log_channel.send(content=content, embed=fallback_embed)
+    
+    async def scan_loop(self):
+        await self.scan_message(await self.scan_queue.get())
+        self.bot.loop.create_task(self.scan_loop())
 
     async def after_invoke_hook(self, ctx: commands.Context):
 
-        await self.scan_message(ctx.message)
+        await self.scan_queue.put(ctx.message)
 
     @commands.Cog.listener()
     async def on_message(self, message: discord.Message):
@@ -284,13 +297,13 @@ class WordWatch(BaseModule):
 
         guild_prefix = await self.bot.command_prefix(self.bot, message)
         if not message.content.startswith(guild_prefix):
-            await self.scan_message(message)
+            await self.scan_queue.put(message)
     
     @commands.Cog.listener()
     async def on_message_edit(self, old: discord.Message, new: discord.Message):
 
         if old.content != new.content:
-            await self.on_message(new)
+            await self.scan_queue.put(new)
 
     @commands.command(name='ww.watch')
     @commands.check_any(commands.has_permissions(administrator=True), has_privlidged_role_check())
@@ -485,15 +498,15 @@ class WordWatch(BaseModule):
 
     @commands.command(name='ww.remove')
     @commands.check_any(commands.has_permissions(administrator=True), has_privlidged_role_check())
-    async def ww_remove(self, ctx: commands.Context, *ranges: str):
+    async def ww_remove(self, ctx: commands.Context, *terms: str):
 
         to_delete = set()
-        for range_ in ranges:
-            if '-' in range_:
-                lower, upper = [int(i) for i in range_.split('-')]
+        for term in terms:
+            if '-' in term:
+                lower, upper = [int(i) for i in term.split('-')]
                 to_delete.update(range(lower, upper+1))
             else:
-                to_delete.add(int(range_))
+                to_delete.add(int(term))
 
         indices = sorted(list(to_delete))
         errors = []
@@ -547,7 +560,7 @@ class WordWatch(BaseModule):
                 mention_type, id = resolve_mention(reference)
             else:
                 mention_type = await self.utils.guess_id(id, ctx.guild)
-            if mention_type in [MentionType.channel, MentionType.role]:
+            if mention_type in [MentionType.channel, MentionType.role, MentionType.user]:
                 if await self.is_id_ignored(ctx.guild.id, id):
                     duplicates += 1
                 else:
@@ -582,10 +595,7 @@ class WordWatch(BaseModule):
 
         errors = []
         for index, reference in enumerate(references):
-            try:
-                id = mention2id(reference, MentionType.channel)
-            except InvalidId:
-                id = mention2id(reference, MentionType.role)
+            id = mention2id(reference)
             try:
                 target_ignore = await WordWatchIgnore.get(target_id=id)
             except DoesNotExist: